Initial try at a .travis.yml file
This migrates the README.md and odbc-test.reb out of the Ren-C repo
and into here, so that it can have a dedicated .travis.yml.

As a first try to get things going, this copies code from the HTTPD
server's test...which clones a Rebol and then tries using it to
run the tests.  There is an OS X matrix entry as well, but that is to
be addressed in later commits.