ODBC Github CI Workflows, Eliminate .travis.yml
Travis ceased supporting open source projects, so we've had
to move.  This is an attempt to preserve the work done thus
far in starting tests for the ODBC extension.

This creates 3 separate workflows for each of the databases
that had some code in the original Travis ODBC test: MySQL,
Sqlite, and Firebird (even though Firebird did not work).

It switches to using the GitHub Action for pulling the latest
built R3...even though that is a work in progress.