Add Workflow for Windows MySQL ODBC Test
This uses the default MySQL 8 which is installed on Windows Runners
as an ODBC source.  Headless install of the MySQL ODBC Connector
.msi file is performed, and PowerShell commands are used to register
the DSN.

Windows restricts the ability to put passwords and usernames in the
DSN database.  So this changes to a common denominator of putting
them in the connection string of the test file.