--leave-connected to test not closing connection
This adds a new switch to the odbc test which will run the test but
not close the database before shutdown.

You shouldn't do it, but it's easy enough to do.  Previously this
was causing a panic because the shutdown GC phase was encountering
stale ODBC objects it could no longer clean up because that part of
the shutdown was after the unloading of the ODBC extension.  A
mitigation was added which makes the ODBC extension shutdown do an
enumeration of outstanding entities pointed to by handles and
"neutralize" them, e.g. make it so the only thing the shutdown
phase does is free their memory.