SYS.UTIL.RESCUE for drop table test
The errors raised by ODBC-EXECUTE are not yet definitional.
They will need to be categorized to which ones should be and
which ones should not.

In the meantime, this uses SYS.UTIL.RESCUE...whose obnoxious
location hints that it's a suboptimal thing to be using and
should only be reached for as a last resort.