Remove sandboxing of tests from recover, outdent
There's no really good reason why the test was being bracketed in a
panic recovery.  The exit status of 0 or 1 should happen naturally
(that is being tested on its own) and the error should be reported.
It just makes the script unnecessarily complex.

(Sadly a simple change like this touches every line.  Source code
should be a graph, but, that's not the game we're playing here.)