`<else>` instead of true for fallthrough branch
TRUE is no longer defined in Ren-C, it's just a WORD! that you test
with TRUE? (to see if it's true or false, panic otherwise).

So if you're doing a fallthrough in a CASE statement and want to have
a truthy thing in the last position, `<else>` is about as good a
choice as anything.