Use ~null~ quasiform for ODBC null
So long as rows are going to be given back as BLOCK!, it can't actually
be a null antiform.

Choosing a quasiform for the Rebol representation makes it easier to
use with REIFY and DEGRADE, to go back and forth from the antiform...
and more likely to cause friction vs. being interpreted simply as
the word "null".